refactor: 영상 분석 상세페이지 경로 및 API 엔드포인트 변경 (#41)
