+++ src/app/videos/[videoId]/page.tsx
@@ -1,4 +1,4 @@
-import Detail from "@components/detail";
+import Detail from "src/components/videos";
 
 export default async function Page({ params }: { params: Promise<{ videoId: string }> }) {
     const { videoId } = await params;
