+++ src/components/videos/WarningBanner.tsx
